Removing subversion, nbehave and ilmerge; Updated Nuget

+++ src/AutoMapperSamples/Mappers/Enumerations.cs
@@ -1,6 +1,6 @@
 using System;
 using AutoMapper;
-using NBehave.Spec.NUnit;
+using Should;
 using NUnit.Framework;
 
 namespace AutoMapperSamples.Mappers
@@ -26,7 +26,7 @@ namespace AutoMapperSamples.Mappers
 			public void Example()
 			{
 				Mapper.Map<OrderStatus, OrderStatusDto>(OrderStatus.InProgress).ShouldEqual(OrderStatusDto.InProgress);
-				Mapper.Map<OrderStatus, short>(OrderStatus.Complete).ShouldEqual(1);
+				Mapper.Map<OrderStatus, short>(OrderStatus.Complete).ShouldEqual((short)1);
 				Mapper.Map<OrderStatus, string>(OrderStatus.Complete).ShouldEqual("Complete");
 				Mapper.Map<short, OrderStatus>(1).ShouldEqual(OrderStatus.Complete);
 				Mapper.Map<string, OrderStatus>("Complete").ShouldEqual(OrderStatus.Complete);
